# backend-go/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.roleBody:
    properties:
      allowed_modem_ids:
        description: nil=不修改，[]uint{}=清除限制
        items:
          type: integer
        type: array
      can_approve_requests:
        type: boolean
      can_support:
        type: boolean
      can_view_history:
        type: boolean
      can_view_sim:
        type: boolean
      description:
        type: string
      name:
        type: string
      read_only:
        type: boolean
    type: object
  handlers.setRolesBody:
    properties:
      role_ids:
        items:
          type: integer
        type: array
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: 多USB 4G Modem 管理系统 API
  title: SimNexus API
  version: "1.0"
paths:
...
  /api/v1/roles/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: 获取角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.roleBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - 角色管理
  /api/v1/roles/{id}:
    delete:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - 角色管理
    patch:
      consumes:
      - application/json
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 修改字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.roleBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 修改角色
      tags:
      - 角色管理
  /api/v1/roles/users/{id}/roles:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.setRolesBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 设置用户角色
      tags:
      - 角色管理
...
securityDefinitions:
  BearerAuth:
    description: '格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"